requirement.txt
fastapi
numpy
uvicorn
pydub
librosa
soundfile
scipy
noisereduce
pedalboard
python-dotenv
websockets
